Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/window-onerror-runtime-error-expected.txt

CONSOLE MESSAGE: line 36: TypeError: undefined is not an object (evaluating 'window.nonexistentproperty.oops')
CONSOLE MESSAGE: line 1515: Error: Tried to create a test with file_is_test
CONSOLE MESSAGE: line 1515: Error: Tried to create a test with file_is_test
CONSOLE MESSAGE: line 1515: Error: Tried to create a test with file_is_test

PASS window.onerror: runtime scripterrors 
